requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.1.7
django-rest-knox==4.2.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
Pillow==9.4.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
